So, the issue I was having was that since I have an event listener for the enter keydown on both the answerInput txt element (to run the check answer function) AND the closeButton on the generated pop up to close the pop up there were problems.
If you hit enter to check the answer, it will run the function to check the answer and generate the popup BUT, since it stays focused on the answerInput text element, if you hit enter to close the pop up it would still be focused on the textinput element and would generate a pop up infinitely. BUT .focus() is a method that switches the focus to any focusable element. SO I started using .focus() between the answerInput and the closeButton.